integers are a pain

+++ src/lib/operators/join_sort_merge/radix_cluster_sort.hpp
@@ -69,12 +69,12 @@ class RadixClusterSort {
   virtual ~RadixClusterSort() = default;
 
   template <typename T2>
-  static std::enable_if_t<std::is_integral_v<T2>, uint32_t> get_radix(T2 value, size_t radix_bitmask) {
+  static std::enable_if_t<std::is_integral_v<T2>, size_t> get_radix(T2 value, size_t radix_bitmask) {
     return static_cast<int64_t>(value) & radix_bitmask;
   }
 
   template <typename T2>
-  static std::enable_if_t<std::is_floating_point_v<T2>, uint32_t> get_radix(T2 value, size_t radix_bitmask) {
+  static std::enable_if_t<std::is_floating_point_v<T2>, size_t> get_radix(T2 value, size_t radix_bitmask) {
     PerformanceWarning("Using hash to perform bit_cast/radix partitioning of floating point number");
     return std::hash<T2>{}(value) & radix_bitmask;
   }
